Move shell to a separate class
